+++ tests/testthat/test-confint.R
@@ -1,3 +1,10 @@
+context("Confidence intervals")
+set.seed(12345)
+
+# CRAN skip atlas check fix
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
+
 # Handling warning messages coming from montecarlo integration
 handler_validmc <- function(msg) {
     if (any(grepl("out of range", msg))) invokeRestart("muffleWarning")