+++ tests/testthat/test-fittingComp.R
@@ -1,6 +1,9 @@
 context("Fitting-Comp risk")
+set.seed(12345)
+
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 
 n <- 100