+++ tests/testthat/test-gam.R
@@ -1,6 +1,9 @@
 context("GAMs")
+set.seed(12345)
+
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 # Skip tests if mgcv is not installed
 testthat::skip_if_not_installed("mgcv")