+++ tests/testthat/test-gbm.R
@@ -1,6 +1,11 @@
 context("GBMs")
+set.seed(12345)
+
+testthat::skip_if(TRUE, message = 'GBM is not implemented')
+
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 # Skip tests if gbm is not installed
 testthat::skip_if_not_installed("gbm")