+++ tests/testthat/test-glmnet.R
@@ -1,7 +1,9 @@
 context("glmnet")
+set.seed(12345)
 
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 # Skip tests if gbm is not installed
 testthat::skip_if_not_installed("glmnet")