+++ tests/testthat/test-matrix.R
@@ -1,7 +1,9 @@
 context("Matrix interface")
-# CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+set.seed(12345)
 
+# CRAN skip atlas check fix
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 N <- 1000; p <- 30
 nzc <- 0.33 * p