+++ tests/testthat/test-popTime.R
@@ -1,7 +1,9 @@
 context("popTime methods")
+set.seed(12345)
 
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 nobs <- 500