+++ tests/testthat/test-sampling.R
@@ -1,7 +1,9 @@
 context("Sampling")
+set.seed(12345)
 
 # CRAN skip atlas check fix
-testthat::skip_if(grepl(pattern = "atlas",sessionInfo()$BLAS,ignore.case=TRUE))
+testthat::skip_if(grepl(pattern = "atlas", sessionInfo()$BLAS,
+                        ignore.case = TRUE))
 
 # Create simulated data with competing risks----
 nobs <- 500